New projects created for solution modules: parser, analyser, main program. Files moved around to match solution structure.
