--- src/QuestionStatistics.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace QuizResults
-{
-	public class QuestionStatistics
-	{
-		public uint RightAnswersAmount { get; set; }
-		public uint WrongAnswersAmount { get; set; }
-	}
-}
