+++ src/QuizData.Parser/Models/Person.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace QuizResults.Models
+namespace QuizData.Parser.Models
 {
 	public class Person
 	{
