+++ src/QuizData/appsettings.json
@@ -1,0 +1,3 @@
+{
+    "DataFilePath": "..\\holyjs.csv"
+}
