--- src/appsettings.json
@@ -1,3 +1,0 @@
-{
-    "DataFilePath": "holyjs.csv"
-}
