+++ test/QuizData.Parser.Test/QuizData.Parser.Test.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>
@@ -14,7 +14,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\QuizResults.csproj" />
+    <ProjectReference Include="..\..\src\QuizData.Parser\QuizData.Parser.csproj" />
   </ItemGroup>
 
 </Project>
